reduce duplication in kubernetes integration sample test
